split s3v2 reader and writer
